pam-extensions: Add new ExternalAuth extension
Sometimes the user needs to complete authentication with an external
identity provider instead of talking to a local authentication service.

That may require the user to e.g. pull out their phone and go to a
URL.

This commit adds a new interface to facilitate that use case. It
provides a way to inform the user about a URI they need to visit.